requirements.txt
PyYAML~=6.0
zserio~=2.12